Fixed user-account logic
