+++ src/main/java/com/epam/lab/paymentsystem/service/AccountService.java
@@ -2,19 +2,46 @@ package com.epam.lab.paymentsystem.service;
 
 import com.epam.lab.paymentsystem.dto.AccountDto;
 import com.epam.lab.paymentsystem.entities.Account;
-import com.epam.lab.paymentsystem.entities.Card;
 import java.util.List;
 
+/**
+ * AccountService is the interface that specifies the service methods.
+ */
 public interface AccountService {
 
+  /**
+   * Returns list of accounts by user's login.
+   *
+   * @param login user's login
+   * @return list of accounts
+   */
   List<Account> getAllAccountsByLogin(String login);
 
+  /**
+   * Saves in the database account entity converted from accountDto.
+   *
+   * @param account account dto passed by controller
+   * @return account entity
+   */
   Account createAccount(AccountDto account);
 
+  /**
+   * Provides creation of the money transfer transaction.
+   *
+   * @param source source account
+   * @param target target account
+   * @param amount amount of transaction
+   * @throws UnsupportedOperationException if there's not enough money on source account
+   *                                       or was passed identical accounts
+   */
   void makeTransaction(Account source, Account target, long amount)
       throws UnsupportedOperationException;
 
+  /**
+   * Returns account by given id.
+   *
+   * @param id id
+   * @return account entity
+   */
   Account getAccountById(long id);
-
-  List<Card> getAllCardsByAccount(Account account);
 }
