+++ src/main/java/com/epam/lab/paymentsystem/service/CardService.java
@@ -11,6 +11,4 @@ public interface CardService {
   List<Card> getAllCardsByLogin(String login);
 
   Card createCard(Card card);
-
-  Account getAccountByCard(Card card);
 }
